Organize Rust relay source
