+++ src/static/dashboard.html
@@ -506,7 +506,7 @@ Content-Type: application/json
           <pre>export LITELLM_GATEWAY_URL=https://gateway.litellm-sandbox.ai
 export LITELLM_GATEWAY_API_KEY=sk-...
 
-./install.sh --set-system-proxy "Wi-Fi"
+./src/install.sh --set-system-proxy "Wi-Fi"
 open http://127.0.0.1:4142/
 
 curl --cacert ~/.litellm-relay/mitm/litellm-relay-ca.pem \
